backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2.3
django-ninja==1.4.3
django-cors-headers==4.3.1
et_xmlfile==2.0.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
